refactor: path,hooks
